requirements-dev.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.10.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.1
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-openai==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
openai==2.5.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
psutil==7.1.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
torch==2.2.2
tqdm==4.67.1
transformers==4.51.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
